+++ back/src/main/java/com/slide/SlideService.java
@@ -60,4 +60,40 @@ public class SlideService {
 		}
 	}
 
+	public void updateAllSlidesLessonName(String oldName, String newName, String unitName, Lesson lesson){
+		List<Slide> slides = this.repository.findByContentContainingAndContentContaining("insert.slide", unitName + "/" + oldName + "/");
+		String newContent;
+
+		for(Slide slide : slides){
+			newContent = slide.getContent().replace(unitName + "/" + oldName + "/", unitName + "/" + newName + "/");
+			slide.setContent(newContent);
+			save(slide);
+		}
+
+		for(Slide slide : lesson.getSlides()){
+			if(slide.getContent().contains("insert.slide") && slide.getContent().contains("/" + oldName + "/")){
+				newContent = slide.getContent().replace(unitName + "/" + oldName + "/", unitName + "/" + newName + "/");
+				slide.setContent(newContent);
+			}
+		}
+	}
+
+	public void updateAllSlidesUnitName(String oldName, String newName){
+		List<Slide> slides = this.repository.findByContentContainingOrContentContaining("insert.card/" + oldName + "/",
+				"insert.slide/" + oldName + "/");
+		String newContent;
+
+		for(Slide slide : slides){
+			if(slide.getContent().contains("insert.card/" + oldName + "/")){
+				newContent = slide.getContent().replace("insert.card/" + oldName + "/", "insert.card/" + newName + "/");
+				slide.setContent(newContent);
+				save(slide);
+			} else {
+				newContent = slide.getContent().replace("insert.slide/" + oldName + "/", "insert.slide/" + newName + "/");
+				slide.setContent(newContent);
+				save(slide);
+			}
+		}
+	}
+
 }