+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Button to flip relations on the unit diagram
 - Icon shows when student has answered a question
 - Itinerary tree keeps opened changing the tab or adding new blocks
+- All slides are updated when changing a lesson, card, unit or slide name
 
 ### Fixed